Refactor get location & weather

+++ app/src/main/java/com/tustar/demo/ui/weather/Weather.kt
@@ -2,6 +2,6 @@ package com.tustar.demo.ui.weather
 
 data class Weather(
     val address: String,
-    val temp: Float,
+    val temp: Int,
     val daily: String
 )
